moves NetworkSafeCaller.kt to data package
