+++ app/src/main/java/com/shevaalex/android/plugev/data/openchargemap/ChargingStationRepositoryImpl.kt
@@ -2,7 +2,7 @@ package com.shevaalex.android.plugev.data.openchargemap
 
 import com.shevaalex.android.plugev.data.openchargemap.network.model.toDomainModel
 import com.shevaalex.android.plugev.data.openchargemap.network.service.ChargingStationRetrofitService
-import com.shevaalex.android.plugev.domain.NetworkSafeCaller
+import com.shevaalex.android.plugev.data.common.network.NetworkSafeCaller
 import com.shevaalex.android.plugev.domain.openchargemap.model.ChargingStation
 import com.shevaalex.android.plugev.domain.openchargemap.model.DataResult
 import com.shevaalex.android.plugev.domain.openchargemap.repository.ChargingStationRepository
